Finished merge from risk management to combines
